# [IMP] base: specify fields to `fields_get` in view postprocessing