--- a/odoo/addons/base/models/ir_ui_view.py
+++ b/odoo/addons/base/models/ir_ui_view.py
@@ -2196,7 +2196,7 @@ class NameManager:
 
     @lazy_property
     def field_info(self):
-        field_info = self.model.fields_get()
+        field_info = self.model.fields_get(attributes=['invisible', 'states', 'readonly', 'required'])
         has_access = functools.partial(self.model.check_access_rights, raise_exception=False)
         if not (has_access('write') or has_access('create')):
             for info in field_info.vals():